requirements.txt
flask
ultralytics
werkzeug
ffmpeg-python
